Move TestListSlackUser and skip
