+++ store/csv_test.go
@@ -7,12 +7,22 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/joho/godotenv"
 	"github.com/wantedly/slack-mention-converter/models"
 )
 
 var (
 	tmpdir string
 )
+
+func loadEnv() error {
+	err := godotenv.Load(".env")
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
 
 func sameUser(u1, u2 *models.User) bool {
 	return u1.LoginName == u2.LoginName && u1.SlackName == u2.SlackName
@@ -76,6 +86,27 @@ func TestUserFlow(t *testing.T) {
 	if !sameUser(user, awakia) {
 		t.Errorf("user should be :%#v, but: %#v", awakia, user)
 	}
+}
+
+// pending
+func TestListSlackUser(t *testing.T) {
+	t.Skip("TestListSlackUser is pending")
+
+	csv := NewCSV(tmpdir)
+
+	if err := loadEnv(); err != nil {
+		t.Fatalf(err.Error())
+	}
+
+	slackUsers, err := csv.ListSlackUsers()
+	if err != nil {
+		t.Errorf("%v", err)
+	}
+	if len(slackUsers) < 10 {
+		t.Fatalf("Data is less than 10, expected a much more larger number")
+	}
+	t.Logf("%v", slackUsers[0].ID)
+	t.Logf("%v", slackUsers[0].Name)
 }
 
 func TestMain(m *testing.M) {
